Bug fix so clone.py won't delete all files in a dir So I was originally using the /tmp folder to store all the garbage files, then just deleting everything in there. While that works just fine for /tmp, since I'm using the home directory by default, it was bad news. I was also accidentally compressing everything in the output directory, so since by default it uses home, I was compressing the entire home directory. Yikes.
Also fixed some stuff with how the program checks if certain flags are on (-h, clone, etc.) to where it just loops through all the arguments and checks where and if they exist. If they're put in a certain incorrect order (like putting flatpak-clone clone -h since that would compress an app called -h) it gives an error.